+++ bin/modules/aka_log.py
@@ -15,11 +15,15 @@
 import logging
 
 
-def init(loglevel='WARNING', loggername=None):
+def init(loglevel='WARNING', loggername=None, jsonlogs: bool = False):
     global log
     log = logging.getLogger(loggername)
     console_handler = logging.StreamHandler()
-    formatter = logging.Formatter('%(asctime)s %(name)s %(levelname).1s %(message)s')
+
+    if jsonlogs:
+        formatter = logging.Formatter('{"timestamp": "%(asctime)s", "log_level": "%(levelname)s", "component": "%(name)s" "message": "%(message)s"}')
+    else:
+        formatter = logging.Formatter('%(asctime)s %(name)s %(levelname)s %(message)s')
     console_handler.setFormatter(formatter)
     log.addHandler(console_handler)
     log.setLevel(loglevel)