+++ bin/uls_config/global_config.py
@@ -50,10 +50,11 @@ output_choices = ['TCP', 'HTTP', 'UDP', 'RAW', 'FILE']         # Definition of O
 transformation_choices = ['MCAS', 'JMESPATH']
 
 
-    # LogLevels
+    # ULS Logging & LogLevels
 log_levels_available = ['DEBUG', 'INFO', 'WARNING', 'ERROR', 'CRITICAL']
 log_level_default = 'WARNING'
 log_debugloglines_default = False
+log_jsonlog = False
 
 # INPUT Configuration
 input_rerun_retries = 3                         # Number of rerun attempts before giving up