+++ docs/CHANGELOG.md
@@ -9,6 +9,8 @@
 
 - **Features**
 
+- **Minor improvements**
+
 - **Bugfixes**
   - issue when using jmespath transformation the result was not proper json - big thanks to @bart-parka for coming up with a PR to fix this