+++ docs/CHANGELOG.md
@@ -8,6 +8,7 @@
 |Author|mschiess@akamai.com
 
 - **Features**
+  - Prometheus monitoring support added to allow smoother monitoring into third party (prometheus compatible) monitoring sytems. More information [here](MONITORING.md#prometheus)
   - CallHome (opt-out) function to enable the ULS team to collect anonymous statistics & usage information - more information [here](./FAQ.md#i-do-not-want-to-send-any-data-to-akamai)
 
 - **Minor improvements**