+++ docs/CHANGELOG.md
@@ -11,6 +11,7 @@
   - Prometheus monitoring support added to allow smoother monitoring into third party (prometheus compatible) monitoring sytems. More information [here](MONITORING.md#prometheus)
   - CallHome (opt-out) function to enable the ULS team to collect anonymous statistics & usage information - more information [here](./FAQ.md#i-do-not-want-to-send-any-data-to-akamai)
   - Added the option to toggle Log Output towards "JSON" format ([see feature request](https://github.com/akamai/uls/issues/66))
+  - Added an option to manipulate the ULS internal logging format ([see feature request](https://github.com/akamai/uls/issues/66) and the [Additional Features section](ADDITIONAL_FEATURES.md#uls-logformat))
   
 
 - **Minor improvements**